+++ CHANGES.md
@@ -1,5 +1,10 @@
 #Changes
 
+## Unreleased
+
+* Fix bug when trying to use XOAUTH to authenticate. - Ben Hamill
+* Remove several cases of excessive argument type checking. - Ben Hamill
+
 ## 1.1.0
 
 * Added `:login` authentication method to `Luggage.new`. - Ben Hamill